+++ prettyprinter/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.0.1 (Next release)
+
+- Rendering directly to a handle is now more efficient in the `Text` renderer,
+  since no intermediate `Text` is generated anymore.
+
 # 1.1
 
 - Allow `alterAnnotations` to convert one annotation to multiple ones, to